JP-23 Final Changes to this feature
Change the manual drawn "OH" ball to a fillText method. Added visual
debug boxes. Corrected the collision boxes of everything. However, with
the "N" letter there is an offset needed to make sure the collision
with the ball lines up correctly. It's on line 130 of index.js. Not sure
why that offset of X pixels was needed. But added it just to make this
feature functional.